Move endless and infinite code fully into dev
It is not useful long-term except insofar as a starting point for
new PRs that do it correctly, and to run demos in the meantime.

Some was already in dev; the portions that were non-dev were too
optimistic that they would need future needs, and were overly
coupled to demo-only considerations.
