+++ drake/automotive/dev/endless_road_car_config.named_vector
