+++ drake/automotive/dev/endless_road_car_state.named_vector
