+++ drake/automotive/dev/endless_road_oracle.cc
@@ -10,8 +10,8 @@
 
 #include <Eigen/Geometry>
 
+#include "drake/automotive/dev/infinite_circuit_road.h"
 #include "drake/automotive/maliput/api/lane_data.h"
-#include "drake/automotive/maliput/utility/infinite_circuit_road.h"
 #include "drake/common/drake_assert.h"
 #include "drake/common/text_logging.h"
 #include "drake/systems/framework/vector_base.h"
