+++ drake/automotive/dev/endless_road_oracle_output.named_vector
