+++ drake/automotive/dev/infinite_circuit_road.h
