+++ drake/automotive/dev/test/endless_road_oracle_test.cc
@@ -5,8 +5,8 @@
 
 #include "gtest/gtest.h"
 
+#include "drake/automotive/dev/infinite_circuit_road.h"
 #include "drake/automotive/maliput/monolane/builder.h"
-#include "drake/automotive/maliput/utility/infinite_circuit_road.h"
 
 namespace drake {
 namespace automotive {
